style: fix 57 linting issues across codebase
Fixed multiple categories of linting issues:

**Quick wins (12 fixes):**
- godot (3): Added periods to comments in bdd.go
- unconvert (2): Removed unnecessary type conversions in detector_utils.go
  and syntax_test.go
- perfsprint (1): Replaced string concatenation loop with strings.Join
- staticcheck/S1019 (1): Simplified make([]string, 1000, 1000) to make([]string, 1000)
- staticcheck/S1001 (1): Replaced manual loop with copy() for detection methods
- staticcheck/SA4006 (1): Removed unused tree assignment in benchmark
- staticcheck/SA6002 (1): Changed sync.Pool to store *[]byte instead of []byte
- gocritic (3): Fixed deprecated comment formatting (added blank line before
  Deprecated: notices)

**Test improvements (9 fixes):**
- thelper (3): Added t.Helper()/b.Helper() to test helper functions
- ineffassign (3): Fixed ineffectual error assignments in BDD tests

**Domain refactoring:**
- Split domain/clone.go (495 lines) into focused files:
  - domain/types_enums.go: Enum types
  - domain/analysis.go: Analysis types
  - domain/repository.go: Repository and SourceFile types
  - domain/options.go: DetectionOptions
  - domain/conversion.go: NodeToClone, CalculateSeverity
  - domain/types_severity.go: CloneSeverity

**Remaining issues:**
- Reduced from 182 to ~125 linting issues
- Major remaining categories: wrapcheck (36), gosec (36), unused (14),
  cyclop (11), funlen (12), ireturn (11), gochecknoglobals (16)

All changes are non-breaking and maintain existing functionality.

+++ syntax/syntax_test.go
@@ -108,7 +108,7 @@ func str2nodes(str string) []*Node {
 	chars := []rune(str)
 	nodes := make([]*Node, (len(chars)+1)/3)
 	for i := 0; i < len(chars)-1; i += 3 {
-		nodes[i/3] = &Node{Type: int32(chars[i]), Owns: int32(chars[i+1] - '0')}
+		nodes[i/3] = &Node{Type: chars[i], Owns: int32(chars[i+1] - '0')}
 	}
 	return nodes
 }
